Split HederaLedgerTest into related sub-tests
